requirements.txt
# Google Ads Dashboard Requirements - Optimized for Vercel
# Core dependencies for the Streamlit Google Ads Dashboard

# Streamlit and UI
streamlit==1.28.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
plotly==5.17.0
altair==5.1.0

# Google Ads API
google-ads==22.1.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data Processing
pandas==2.1.0
numpy==1.24.0
pytz==2023.3

# Configuration and Utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# HTTP and API
requests==2.31.0

# AI Ad Generation Dependencies
openai==1.3.0

# Essential features only
openpyxl==3.1.2